backend/requirements.txt
asgiref==3.8.1
dj-database-url==2.3.0
Django==5.2.1
django-cors-headers==4.7.0
djangorestframework==3.16.0
environs==14.1.1
gunicorn==23.0.0
marshmallow==3.26.1
packaging==24.2
psycopg2-binary==2.9.10
python-dotenv==1.1.0
sqlparse==0.5.3
typing_extensions==4.13.1
tzdata==2025.2